+++ src/lib.rs
@@ -1,8 +1,5 @@
 mod config;
 mod controller;
-mod converter;
-mod fluentd;
-mod matcher;
 mod operation_log;
 mod producer;
 mod report;
@@ -10,8 +7,5 @@ mod zeek;
 
 pub use config::{Config, InputType, OutputType};
 pub use controller::Controller;
-pub use converter::Converter;
-pub use fluentd::SizedForwardMode;
-pub use matcher::Matcher;
 pub use producer::Producer;
 pub use report::Report;